% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tirar_acentos.R
\name{tirar_acentos}
\alias{tirar_acentos}
\title{Tirar acentos}
\usage{
tirar_acentos(texto)
}
\arguments{
\item{texto}{texto para tirar acentos}
}
\value{
texto sem acentos
}
\description{
Tirar acentos
}
